Prevent deletion of report request entry if amazon processing status is DoneNoData. Added unit test.

+++ src/EasyMWS/EasyMWS/Processors/RequestReportProcessor.cs
@@ -202,13 +202,23 @@ namespace MountainWarehouse.EasyMWS.Processors
 				}
 				else if (reportGenerationInfo.ReportProcessingStatus == AmazonReportProcessingStatus.DoneNoData)
 				{
-					reportRequestService.Delete(reportRequestEntry);
-					_logger.Warn($"{genericProcessingInfo}. The Report request entry will now be removed from queue.");
+                    if (_options.InvokeCallbackForReportStatusDoneNoData)
+                    {
+                        reportRequestEntry.ReportProcessRetryCount = 0;
+                        reportRequestService.Update(reportRequestEntry);
+                        _logger.Warn($"{genericProcessingInfo}. The Report was successfully processed by Amazon, but it contains no data. An callback invocation will be done with a null stream argument.");
+                    }
+                    else
+                    {
+                        reportRequestService.Delete(reportRequestEntry);
+                        _logger.Warn($"{genericProcessingInfo}. The Report request entry will now be removed from queue.");
+                    }
 				}
 				else if (reportGenerationInfo.ReportProcessingStatus == AmazonReportProcessingStatus.Submitted 
 					  || reportGenerationInfo.ReportProcessingStatus == AmazonReportProcessingStatus.InProgress)
 				{
-					_logger.Info($"{genericProcessingInfo}. The report processing status will be checked again at the next poll request.");
+                    reportRequestService.Update(reportRequestEntry);
+                    _logger.Info($"{genericProcessingInfo}. The report processing status will be checked again at the next poll request.");
 				}
 				else if (reportGenerationInfo.ReportProcessingStatus == AmazonReportProcessingStatus.Cancelled)
 				{
